Renaming and refactoring of modules and components for bettern understanding
