+++ src/components/DocEditor.vue
