+++ src/views/Index.vue
@@ -2,10 +2,10 @@
   <div
     refs="docPlace"
     id="app"
-    class="content-center w-1/3 m-auto text-center items-center  flex-col"
+    class="content-center w-1/3 m-auto text-center items-center flex-col"
   >
-  <h1>HERE I AM DUMPING ALL COMPONENTS TO TEST THEM</h1>
-    <div id="nav"><router-link to="/">Home</router-link> |</div>
+    <h1>HERE I AM DUMPING ALL COMPONENTS TO TEST THEM</h1>
+    <div id="nav"><router-link to="/">Home</router-link>|</div>
     <div>
       <h1 style="color:red" class="bg-gray-25">{{ query }}</h1>
       <button class="primary-button" @click="gqlQuery()">
@@ -20,27 +20,26 @@
         <br />
         <br />
         <DocsContents></DocsContents>
-        <OpenFolder></OpenFolder>
+        <OpenDocs></OpenDocs>
       </div>
       <!-- <h1>What an Doc</h1> -->
-      <Doc id="1"></Doc>
+      <DocEditor id="1"></DocEditor>
     </div>
   </div>
 </template>
 
 <script>
 import Vue from "vue";
-import Doc from "@/components/Doc";
-import OpenFolder from "@/components/OpenFolder";
+import DocEditor from "@/components/DocEditor";
+import OpenDocs from "@/components/OpenDocs";
 import DocsServices from "@/services/index";
 import DocsContents from "@/components/DocsContents";
 
-
 // import DataCell from "~/components/DataCell";
-// import OpenFolder from "~/components/OpenFolder";
+// import OpenDocs from "~/components/OpenDocs";
 
 export default {
-  components: { Doc, OpenFolder, DocsContents },
+  components: { DocEditor, OpenDocs, DocsContents },
   data: function() {
     return {
       path: "",
@@ -55,7 +54,7 @@ export default {
       this.query = result.data.cwd;
     },
     addDoc() {
-      let ComponentClass = Vue.extend(Doc);
+      let ComponentClass = Vue.extend(DocEditor);
       let instance = new ComponentClass();
       instance.$mount(); // pass nothing
       this.$refs.docPlace.appendChild(instance.$el);
